attemp2

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-    float calc = toBeDivided / 5;
+    double calc = toBeDivided / 5.0;
         System.out.println(calc);
     }
 }
